+++ Andy/tzudoku but red detect lol.swiftpm/.swiftpm/playgrounds/CachedManifest.plist
