+++ Andy/tzudoku but red detect lol.swiftpm/.swiftpm/playgrounds/DocumentThumbnail.png
