+++ Andy/tzudoku but red detect lol.swiftpm/Package.swift
